Separate out CPP interface and implementation
